+++ Modules/Administration/Exceptions/AdministrationException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Modules\Administration\Exceptions;
+
+use RuntimeException;
+
+class AdministrationException extends RuntimeException
+{
+}