+++ Modules/Administration/Exceptions/MonitoringException.php
@@ -2,8 +2,6 @@
 
 namespace Modules\Administration\Exceptions;
 
-use RuntimeException;
-
-final class MonitoringException extends RuntimeException
+final class MonitoringException extends AdministrationException
 {
 }