+++ Modules/Administration/Exceptions/SecurityPolicyException.php
@@ -2,8 +2,6 @@
 
 namespace Modules\Administration\Exceptions;
 
-use RuntimeException;
-
-final class SecurityPolicyException extends RuntimeException
+final class SecurityPolicyException extends AdministrationException
 {
 }